+++ src/Silverback.Integration/Messaging/Broker/IBrokerCollection.cs
@@ -13,25 +13,37 @@ namespace Silverback.Messaging.Broker;
 public interface IBrokerCollection : IReadOnlyList<IBroker>
 {
     /// <summary>
-    ///     Returns an <see cref="IProducer" /> to be used to produce to the specified endpoint.
+    ///     Returns an <see cref="IProducer" /> with the specified configuration.
     /// </summary>
     /// <param name="configuration">
     ///     The producer configuration.
     /// </param>
     /// <returns>
-    ///     The <see cref="IProducer" /> for the specified endpoint.
+    ///     A <see cref="Task{TResult}" /> representing the asynchronous operation. The task result contains the
+    ///     <see cref="IProducer" /> with the specified configuration.
+    /// </returns>
+    Task<IProducer> GetProducerAsync(ProducerConfiguration configuration);
+
+    /// <summary>
+    ///     Returns an <see cref="IProducer" /> with the specified configuration.
+    /// </summary>
+    /// <param name="configuration">
+    ///     The producer configuration.
+    /// </param>
+    /// <returns>
+    ///     The <see cref="IProducer" /> with the specified configuration.
     /// </returns>
     IProducer GetProducer(ProducerConfiguration configuration);
 
     /// <summary>
-    ///     Adds an <see cref="IConsumer" /> that will consume from the specified endpoint as soon as the broker
-    ///     is connected. The received messages will be forwarded to the specified callback delegate.
+    ///     Adds an <see cref="IConsumer" /> with the specified configuration that will start consuming as soon as the broker
+    ///     is connected. The received messages will sent through the behaviors pipeline and forwarded to the subscribers.
     /// </summary>
     /// <param name="configuration">
     ///     The consumer configuration.
     /// </param>
     /// <returns>
-    ///     The <see cref="IConsumer" /> for the specified endpoint.
+    ///     The <see cref="IConsumer" /> with the specified configuration.
     /// </returns>
     IConsumer AddConsumer(ConsumerConfiguration configuration);