+++ src/Silverback.Integration/Messaging/Outbound/TransactionalOutbox/OutboxProduceStrategy.cs
@@ -2,6 +2,7 @@
 // This code is licensed under MIT license (see LICENSE file for details)
 
 using System;
+using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Silverback.Diagnostics;
@@ -48,6 +49,8 @@ public class OutboxProduceStrategy : IProduceStrategy, IEquatable<OutboxProduceS
 
         private readonly IOutboundLogger<OutboxProduceStrategy> _logger;
 
+        private readonly SemaphoreSlim _producerInitSemaphore = new(1, 1);
+
         private IProducer? _producer;
 
         public OutboxProduceStrategyImplementation(
@@ -58,15 +61,29 @@ public class OutboxProduceStrategy : IProduceStrategy, IEquatable<OutboxProduceS
             _logger = logger;
         }
 
-        public Task ProduceAsync(IOutboundEnvelope envelope)
+        public async Task ProduceAsync(IOutboundEnvelope envelope)
         {
             Check.NotNull(envelope, nameof(envelope));
 
             _logger.LogWrittenToOutbox(envelope);
 
-            _producer ??= _outboundQueueBroker.GetProducer(envelope.Endpoint.Configuration);
+            _producer ??= await InitProducerAsync(envelope.Endpoint.Configuration).ConfigureAwait(false);
 
-            return _producer.ProduceAsync(envelope);
+            await _producer.ProduceAsync(envelope).ConfigureAwait(false);
+        }
+
+        private async Task<IProducer> InitProducerAsync(ProducerConfiguration configuration)
+        {
+            await _producerInitSemaphore.WaitAsync().ConfigureAwait(false);
+            try
+            {
+                _producer = await _outboundQueueBroker.GetProducerAsync(configuration).ConfigureAwait(false);
+                return _producer;
+            }
+            finally
+            {
+                _producerInitSemaphore.Release();
+            }
         }
     }
 }